// web/src/lib/types/generated/LogicalCondition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LeafCondition } from "./LeafCondition";
import type { LogicalOperator } from "./LogicalOperator";

export type LogicalCondition = { op: LogicalOperator, 
/**
 * In TypeScript, Condition = LogicalCondition | LeafCondition
 */
children: Array<LogicalCondition | LeafCondition>, };